Singleton Views
Forces to have only one instance in the views, in this way when pressing a button several times they do not overlap nor generate a stack of Views. In addition, when closing the Main view all the views that the Singleton contains are closed. The messenger pattern has been discontinued, but it is still active for its future usefulness.